// src/types/users.d.ts
export interface CurrentUser {
    id: string
    avatar: string
    handle: string
    username: string
    is_admin: string
}

export interface UserMetadata {
    avatar_url: string
    email: string
    full_name: string
    name: string
    picture: string
    prefered_username: string
    user_name: string
}